requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==2.1.0
Click==7.0
colorama==0.4.1
Flask==1.0.2
gast==0.2.0
grpcio==1.17.1
h5py==2.9.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
numpy==1.15.4
opencv-python==4.0.0.21
pep8==1.7.1
Pillow==5.3.0
protobuf==3.6.1
pylint==2.2.2
PyYAML==3.13
scipy==1.2.0
six==1.12.0
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
typed-ast==1.1.1
Werkzeug==0.14.1
wrapt==1.10.11
